+++ src/pages/accomodation/registration/assign-rooms.tsx
