+++ src/pages/accomodation/registration/register-destination.tsx
